requirements.txt
# 解析 init.yaml，无标准库替代
PyYAML>=6.0
